+++ test/unit/fuzz-tools.test.js
@@ -282,6 +282,7 @@ import * as isisV302 from '../../lib/isis-v302.js';
 import * as anaphylaxisV303 from '../../lib/anaphylaxis-v303.js';
 import * as dstV304 from '../../lib/dst-v304.js';
 import * as crsV305 from '../../lib/crs-v305.js';
+import * as icansV306 from '../../lib/icans-v306.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -509,6 +510,7 @@ const MODULES = {
   'anaphylaxis-v303.js': anaphylaxisV303,
   'dst-v304.js': dstV304,
   'crs-v305.js': crsV305,
+  'icans-v306.js': icansV306,
   'tb-testing.js': tbTesting,
 };